requirements.txt
crewai
anthropic
openai
python-dotenv
langchain-anthropic
'crewai[tools]'